Finished implementation of BoxLosses. Needs testing.

+++ blond/physics/losses.py
@@ -1,6 +1,9 @@
 from __future__ import annotations
 
+from abc import ABC
 from typing import TYPE_CHECKING
+
+import numpy as np
 
 if TYPE_CHECKING:  # pragma: no cover
     from typing import Any
@@ -11,23 +14,103 @@ from .._core.beam.base import BeamBaseClass
 from .._core.simulation.simulation import Simulation
 
 
-class LossesBaseClass(BeamPhysicsRelevant):
-    def __init__(self) -> None:
+class LossesBaseClass(BeamPhysicsRelevant, ABC):
+    """Base class for labeling/removing lost particles.
+
+    Parameters
+    ----------
+    purge_flagged_macroparticles
+        If true, particles will be immediately removed
+        from the ``Beam`` array when ``track(...)`` is executed.
+
+        If false, the ``Beam.flags`` will be set, but particles will still
+        be considered for beam physics.
+
+    Attributes
+    ----------
+    purge_flagged_macroparticles
+        If true, particles will be immediately removed
+        from the ``Beam`` array when ``track(...)`` is executed.
+
+        If false, the ``Beam.flags`` will be set, but particles will still
+        be considered for beam physics.
+    """
+
+    def __init__(self, purge_flagged_macroparticles: bool) -> None:
         super().__init__()
+        self.purge_flagged_macroparticles = purge_flagged_macroparticles
 
     def track(self, beam: BeamBaseClass) -> None:
-        beam.purge_flagged_entries()
+        """Main simulation routine to be called in the mainloop.
+
+        Parameters
+        ----------
+        beam
+            Beam class to interact with this element
+        """
+        if self.purge_flagged_macroparticles:
+            beam.purge_flagged_entries()
 
 
 class BoxLosses(LossesBaseClass):
+    """Particles outside a rectangle will be flagged lost.
+
+    Parameters
+    ----------
+    purge_flagged_macroparticles
+        If true, particles will be immediately removed
+        from the ``Beam`` array when ``track(...)`` is executed.
+
+        If false, the ``Beam.flags`` will be set, but particles will still
+        be considered for beam physics.
+    t_min
+        Macro-particles with ``dt < t_min`` will be labeled/removed, in [s],
+    t_max
+        Macro-particles with ``dt > t_max`` will be labeled/removed, in [s],
+    e_min
+        Macro-particles with ``dE < t_min`` will be labeled/removed, in [s],
+    e_max
+        Macro-particles with ``dE > t_min`` will be labeled/removed, in [s],
+
+    Attributes
+    ----------
+    t_min
+        Macro-particles with ``dt < t_min`` will be labeled/removed, in [s],
+    t_max
+        Macro-particles with ``dt > t_max`` will be labeled/removed, in [s],
+    e_min
+        Macro-particles with ``dE < t_min`` will be labeled/removed, in [s],
+    e_max
+        Macro-particles with ``dE > t_min`` will be labeled/removed, in [s],
+    """
+
     def __init__(
         self,
+        purge_flagged_macroparticles: bool,
         t_min: backend.float | None = None,
         t_max: backend.float | None = None,
         e_min: backend.float | None = None,
         e_max: backend.float | None = None,
     ) -> None:
-        super().__init__()
+        super().__init__(
+            purge_flagged_macroparticles=purge_flagged_macroparticles,
+        )
+        if t_min is None:
+            # USe float instead of None
+            # for easier implementation of kernels.
+            t_min = np.finfo(backend.float).min
+        if t_max is None:
+            # USe float instead of None
+            # for easier implementation of kernels.
+            t_max = np.finfo(backend.float).max
+        if e_min is None:
+            # USe float instead of None
+            # for easier implementation of kernels.
+            e_min = np.finfo(backend.float).min
+        if e_max is None:
+            # USe float instead of None
+            # for easier implementation of kernels.
+            e_max = np.finfo(backend.float).max
 
         self.t_min = backend.float(t_min)
         self.t_max = backend.float(t_max)
@@ -50,6 +133,17 @@ class BoxLosses(LossesBaseClass):
         turn_i_init: int,
         **kwargs: dict[str, Any],
     ) -> None:
+        """Lateinit method when `simulation.run_simulation` is called.
+
+        simulation
+            Simulation context manager
+        beam
+            Simulation beam object
+        n_turns
+            Number of turns to simulate
+        turn_i_init
+            Initial turn to execute simulation
+        """
         pass
 
     def track(self, beam: BeamBaseClass) -> None:
@@ -60,46 +154,11 @@ class BoxLosses(LossesBaseClass):
         beam
             Beam class to interact with this element
         """
-        backend.loss_box(
+        backend.specials.loss_box(
             beam.write_partial_flags(),
             self.t_min,
             self.t_max,
             self.e_min,
             self.e_max,
         )
-        super().track(beam=beam)
-
-
-class SeparatrixLosses(LossesBaseClass):
-    def __init__(self) -> None:
-        super().__init__()
-        self._simulation: Simulation | None = None
-
-    def on_init_simulation(self, simulation: Simulation) -> None:
-        """Lateinit method when `simulation.__init__` is called.
-
-        simulation
-            Simulation context manager
-        """
-        self._simulation = simulation
-
-    def on_run_simulation(
-        self,
-        simulation: Simulation,
-        beam: BeamBaseClass,
-        n_turns: int,
-        turn_i_init: int,
-        **kwargs: dict[str, Any],
-    ) -> None:
-        pass
-
-    def track(self, beam: BeamBaseClass) -> None:
-        """Main simulation routine to be called in the mainloop.
-
-        Parameters
-        ----------
-        beam
-            Beam class to interact with this element
-        """
-        self._simulation.get_separatrix()  # TODO
         super().track(beam=beam)
